refactor: #140 cursor-colors 파일 명 수정
