+++ apps/client/src/features/code-editor-sidebar/model/cursor-colors.utils.ts
